Check if csv exists before deleting

+++ convert_xlsx_to_csv.ps1
@@ -12,7 +12,9 @@ Get-ChildItem $path -Filter *.xlsx | Foreach-Object{
     'processing ' + $_.FullName
     $wb = $excel.Workbooks.Open($_.FullName)
     $dst_file = $path + "\" + $_.BaseName + ".csv"
-    Remove-Item –path $dst_file
+    if([System.IO.File]::Exists($dst_file)){
+        Remove-Item –path $dst_file
+    }
     $wb.SaveAs($dst_file, 42)  # 42 is Tab delimited UTF-16LE with leading Byte Order Mark
     'saved ' + $dst_file
     $wb.Close($True)
